added specific authorships module
