+++ src/js/modules/articles/actionTypes.js
@@ -5,7 +5,4 @@ export const FETCH_ARTICLES_FAILED = 'forms/FETCH_ARTICLES_FAILED'
 export const DELETE_ARTICLES_REQUESTED = "forms/DELETE_ARTICLES_REQUESTED";
 export const DELETE_ARTICLES_SUCCEEDED = 'forms/DELETE_ARTICLES_SUCCEEDED';
 export const DELETE_ARTICLES_FAILED = 'forms/DELETE_ARTICLES_FAILED'
-export const FETCH_AUTHORSHIPS_REQUESTED = "forms/FETCH_AUTHORSHIPS_REQUESTED";
-export const FETCH_AUTHORSHIPS_SUCCEEDED = 'forms/FETCH_AUTHORSHIPS_SUCCEEDED';
-export const FETCH_AUTHORSHIPS_FAILED = 'forms/FETCH_AUTHORSHIPS_FAILED'
