+++ src/js/modules/articles/selectors.js
@@ -2,15 +2,12 @@
  * Created by nicholas on 7/25/17.
  */
 import { createSelector } from 'reselect'
-import { usersSelector } from '../users/selectors'
 
 export const articlesSelector = state => state.articles.list
-export const authorshipsSelector = state => state.articles.authorships
 
 export const articlesPreviewSelector = createSelector(
   articlesSelector,
-  authorshipsSelector,
-  (articles, authorships) => (
+  articles => (
     articles.map(
       article => {
         const content = article.content
@@ -40,29 +37,3 @@ export const randomArticleSelector = createSelector(
   )
 )
 
-// Returns object with the
-export const contributorsByArticle = createSelector(
-  authorshipsSelector,
-  articlesSelector,
-  usersSelector,
-  (authorships, articles, users) => {
-    let contributors = {}
-    // For every article...
-    articles.map(article => {
-      // Find the authorships corresponding to that article
-      const articleAuthorships = authorships.filter(
-        authorship => authorship.articleId === article.id
-      )
-      // Then find the users in those authorships
-      const articleContributors = articleAuthorships.map(authorship =>
-        users.find(
-          user => authorship.userId === user.id
-        )
-      )
-      // Finally add it to an object under the articleId
-      // Jesus, this is inefficient. Thank god for memoizatio
-      contributors[article.id] = articleContributors
-    })
-    return contributors;
-  }
-)
