+++ src/js/modules/authorships/actionTypes.js
@@ -1,0 +1,6 @@
+/**
+ * Created by nicholas on 8/4/17.
+ */
+export const FETCH_AUTHORSHIPS_REQUESTED = "forms/FETCH_AUTHORSHIPS_REQUESTED";
+export const FETCH_AUTHORSHIPS_SUCCEEDED = 'forms/FETCH_AUTHORSHIPS_SUCCEEDED';
+export const FETCH_AUTHORSHIPS_FAILED = 'forms/FETCH_AUTHORSHIPS_FAILED'
