+++ src/js/modules/authorships/constants.js
@@ -1,0 +1,4 @@
+export default
+{
+    NAME :  'authorships'
+}
