+++ src/js/modules/authorships/index.js
@@ -1,0 +1,4 @@
+import constants from './constants'
+import reducer   from './reducer'
+
+export default { constants, reducer }
