+++ src/js/reducers.js
@@ -5,13 +5,15 @@ import users               from './modules/users'
 import sections            from './modules/sections'
 import forms               from './modules/forms'
 import main                from './modules/main'
+import authorships         from './modules/authorships'
 
 export default combineReducers(
 {
   [articles.constants.NAME]           : articles.reducer,
-  [users.constants.NAME]              : users.reducer,
-  [sections.constants.NAME]           : sections.reducer,
+  [authorships.constants.NAME]        : authorships.reducer,
   [forms.constants.NAME]              : forms.reducer,
   [main.constants.NAME]               : main.reducer,
+  [sections.constants.NAME]           : sections.reducer,
+  [users.constants.NAME]              : users.reducer,
   router                              : routerReducer
 });
